Update and rename WheelTest2.kt to WheelTest.kt
